requirements.txt
pandas
faker
numpy
psycopg2-binary
sqlalchemy
jupyter